converted vanilla JS to stimulus JS and got the first phase of the HP working
