+++ app/views/pages/home.html.erb
@@ -1,3 +1,6 @@
- <div id="homepagebody" data-controller="home">
-  <h1 data-controller="home" id="homepageh1" data-home-value="FantasyFlights">...FantasyFlights</h1>
+ <div id="homepagebody" data-controller="home-page" data-home-page-title-value="...FantasyFlights">
+  <h1 data-home-page-target="hyperplexed" data-action="mouseover->home-page#fire" id="homepageh1" >...FantasyFlights</h1>
+  <div class="d-none" data-home-page-target="reveal">
+    <h6>Stuff to hide</h6>
+  </div>
 </div>
